abstracted all dispatches and selectors into the new hero.service "store facade"
