+++ src/client/app/app.module.ts
@@ -4,13 +4,14 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
-import { HeroService } from './hero.service';
+import { HeroDataService } from './hero-data.service';
 import { HeroListComponent } from './hero-list.component';
 import { HeroDetailComponent } from './hero-detail.component';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import { HeroEffects } from './hero.effects';
 import { reducers } from './hero.reducer';
+import { HeroService } from './hero.service';
 
 @NgModule({
   declarations: [AppComponent, HeroListComponent, HeroDetailComponent],
@@ -22,7 +23,7 @@ import { reducers } from './hero.reducer';
     // StoreModule.forRoot({ reducers }), // TODO: this is not working
     EffectsModule.forRoot([HeroEffects])
   ],
-  providers: [HeroService],
+  providers: [HeroDataService, HeroService],
   bootstrap: [AppComponent]
 })
 export class AppModule {}
